Move seed data for stateful components into mappings folder
