+++ packages/core/src/mappings/FieldSearchBarFull.js
@@ -1,0 +1,57 @@
+import {
+  GROUPS,
+  COMPONENT_TYPES,
+  FORM_TYPES,
+  PROP_TYPES,
+  FIELD_NAME,
+  Triggers,
+} from "@draftbit/types";
+
+export const SEED_DATA = [
+  {
+    name: "Search Bar",
+    tag: "FieldSearchBarFull",
+    description: "A search bar with accompanying search icon and clear button.",
+    category: COMPONENT_TYPES.input,
+    preview_image_url: "{CLOUDINARY_URL}/Field_SearchBar_Full.png",
+    supports_list_render: false,
+    triggers: [Triggers.OnChange], // TODO Triggers.OnSubmit for multiple triggers
+    props: {
+      icon: {
+        group: GROUPS.basic,
+        label: "Icon",
+        description: "Left icon to display",
+        formType: FORM_TYPES.icon,
+        propType: PROP_TYPES.ASSET,
+        defaultValue: null,
+        editable: true,
+        required: false,
+      },
+      placeholder: {
+        group: GROUPS.basic,
+        label: "Placeholder",
+        description: "Input placeholder text",
+        formType: FORM_TYPES.string,
+        propType: PROP_TYPES.STRING,
+        defaultValue: "Search for...",
+        editable: true,
+        required: false,
+      },
+      onSubmit: {
+        group: GROUPS.basic,
+        label: "Submit action",
+        description: "Action to execute on submission",
+        editable: true,
+        required: false,
+        formType: FORM_TYPES.action,
+        propType: PROP_TYPES.STRING,
+        defaultValue: null,
+      },
+      fieldName: {
+        ...FIELD_NAME,
+        defaultValue: "searchBarValue",
+      },
+    },
+    layout: {},
+  },
+];
